ANDES-84 #comment zuist-cluster-fits version 0.1.2-SNAPSHOT with new versions of the dependencies. This used FitsImage instead of JSkyFitsImage.
git-svn-id: file:///home/hoijui/src/IsaViz/tmp2/zvtm.svn/code@5145 74e165f3-c697-40a6-91af-552c8936d7fe

+++ zuist-cluster-fits/trunk/src/main/java/fr/inria/zuist/engine/FitsResourceHandler.java
@@ -11,9 +11,13 @@ public class FitsResourceHandler implements ResourceHandler {
     private static final String SM_ID = "sm="; //scale method in params
     private static final String CF_ID = "cf="; //color filter in params
 
+//    public ResourceDescription createResourceDescription(
+//            long x, long y, String id, int zindex, Region region, 
+//            URL resourceURL, boolean sensitivity, Color stroke, String params){
     public ResourceDescription createResourceDescription(
-            long x, long y, String id, int zindex, Region region, 
+            double x, double y, String id, int zindex, Region region, 
             URL resourceURL, boolean sensitivity, Color stroke, String params){
+
         float scaleFactor = 1;
         FitsImage.ScaleMethod scaleMethod = FitsImage.ScaleMethod.ASINH;
         FitsImage.ColorFilter colorFilter = FitsImage.ColorFilter.NOP;
@@ -49,6 +53,8 @@ public class FitsResourceHandler implements ResourceHandler {
                 scaleFactor, scaleMethod, colorFilter);
         region.addObject(desc);
         return desc;             
-            }
+    }
+
+
 }
